jkjk

+++ server.js
@@ -5,11 +5,13 @@ const nodemailer = require("nodemailer");
 app.use(express.json())
 app.use((req, res, next) => {
     console.log("it's custom middleware")
+    console.log("CURRENT STORE:", otpStore);
     req.example = "exampleValue";
     next();
 });
 const cors = require('cors');
 const { Route } = require("lucide-react");
+const { error } = require("console");
 app.use(cors());
 
 
@@ -26,18 +28,38 @@ app.post("/send_otp", async (req, res) => {
         const userExists = usercontent.find(el => el.email === email);
         const code = Math.floor(1000 + Math.random() * 9000)
         console.log(code)
+
+        if (!email) return res.status(400).json({
+            message: "Email is required"
+        });
+
         if (!userExists) {
             usercontent.push({ email, password: "" });
+            // otpStore[email] = code;
             await fs.writeFile("./users.json", JSON.stringify(usercontent, null, 2));
+            res.status(200).json({
+                status: "ok",
+                message: "New user"
+            })
         }
+        if (userExists) {
 
-        otpStore[email] = code;
+            otpStore[email] = code;
+            console.log("OTP GENERATED:", code);
 
-        const stringusercontent = JSON.stringify(usercontent, null, 2)
-        await fs.writeFile("./users.json", stringusercontent)
-        console.log("Code Sent!")
+            const stringusercontent = JSON.stringify(usercontent, null, 2)
+            await fs.writeFile("./users.json", stringusercontent)
+            console.log("Code Sent!")
+            res.status(200).json({
+                status: "not ok",
+                message: "Old User"
+            })
 
-        const Mail_Template = `<!DOCTYPE html>
+            setTimeout(() => {
+                console.log("AFTER 10s:", email, otpStore[email]);
+            }, 10000);
+
+            const Mail_Template = `<!DOCTYPE html>
 <html>
 <head>
   <meta charset="UTF-8">
@@ -118,33 +140,27 @@ app.post("/send_otp", async (req, res) => {
 </html>
 `
 
-        const transport = nodemailer.createTransport({
-            service: "gmail",
-            auth: {
-                user: "codewithvishal001@gmail.com",
-                pass: "gxgo vita sphh glwl"
-            }
-        });
-        transport.sendMail({
-            to: email,
-            from: "codewithvishal001@gmail.com",
-            subject: "Reset Your Password",
-            html: Mail_Template,
-            text: "Don't share this Code with anyone!"
-        })
-            .then(() => {
-                console.log("Mail sent")
+            const transport = nodemailer.createTransport({
+                service: "gmail",
+                auth: {
+                    user: "codewithvishal001@gmail.com",
+                    pass: "gxgo vita sphh glwl"
+                }
+            });
+            transport.sendMail({
+                to: email,
+                from: "codewithvishal001@gmail.com",
+                subject: "Reset Your Password",
+                html: Mail_Template,
+                text: "Don't share this Code with anyone!"
             })
-            .catch((error) => {
-                console.log("Error: ", error.message)
-            })
-
-
-
-        res.status(200).json({
-            message: "Code Sent to your given email"
-        })
-
+                .then(() => {
+                    console.log("Mail sent")
+                })
+                .catch((error) => {
+                    console.log("Error: ", error.message)
+                })
+        }
     } catch (error) {
         console.log("Error : ", error.message);
         res.status(401).json({ message: error.message });
@@ -164,7 +180,7 @@ app.post("/verify_otp", async (req, res) => {
         if (otpStore[email] != code)
             return res.status(400).json({ message: "Wrong OTP" });
 
-        delete otpStore[email];
+        // delete otpStore[email];
 
         const filecontent = await fs.readFile("./users.json", "utf-8")
         const usercontent = JSON.parse(filecontent)
@@ -175,7 +191,6 @@ app.post("/verify_otp", async (req, res) => {
         } else {
             return res.status(200).json({ status: "new_user" });
         }
-        console.log(otpStore[email])
 
     } catch (error) {
         console.log("Error : ", error.message);
@@ -186,23 +201,23 @@ app.post("/verify_otp", async (req, res) => {
 // Route :: 3
 
 app.post("/login", async (req, res) => {
-    
+
     try {
         const { email, password } = req.body;
         const filecontent = await fs.readFile("./users.json", "utf-8")
         const usercontent = JSON.parse(filecontent)
-        
+
         const userExists = usercontent.find(el => el.email === email);
-        
+
         if (!userExists) {
             return res.status(401).json({ message: "User not found" });
         }
         if (userExists.password == password) {
-            return res.status(200).json({message:"Login successfull"})
+            return res.status(200).json({ message: "Login successfull" })
         }
-        else{
+        else {
             res.status(401).json({ message: "Wrong Password" });
-         }
+        }
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
